+++ forge/src/main/java/top/theillusivec4/polymorph/mixin/integration/sophisticatedbackpacks/AccessorCookingLogic.java
@@ -21,15 +21,13 @@
 
 package top.theillusivec4.polymorph.mixin.integration.sophisticatedbackpacks;
 
-import java.util.function.Supplier;
-import net.p3pp3rf1y.sophisticatedbackpacks.upgrades.smelting.SmeltingLogic;
-import net.p3pp3rf1y.sophisticatedbackpacks.upgrades.smelting.SmeltingLogicContainer;
+import net.p3pp3rf1y.sophisticatedbackpacks.upgrades.cooking.CookingLogic;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.gen.Accessor;
 
-@Mixin(SmeltingLogicContainer.class)
-public interface AccessorSmeltingLogicContainer {
+@Mixin(CookingLogic.class)
+public interface AccessorCookingLogic {
 
   @Accessor(remap = false)
-  Supplier<SmeltingLogic> getSupplySmeltingLogic();
+  void setCookingRecipeInitialized(boolean pVal);
 }